pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grid-football-prediction"
version = "1.0.0"
description = "Advanced AI Football Prediction System"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Football Analytics Team"}
]
keywords = ["football", "nfl", "ncaa", "prediction", "analytics", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "torch>=2.1.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.1.0",
    "optuna>=3.4.0",
    "shap>=0.43.0",
    "pyarrow>=14.0.0",
    "duckdb>=0.9.0",
    "sqlalchemy>=2.0.0",
    "apscheduler>=3.10.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.0",
    "great-expectations>=0.18.0",
    "onnxruntime>=1.16.0",
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.0",
    "torch-geometric>=2.4.0",
    "statsmodels>=0.14.0",
    "plotly>=5.17.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "joblib>=1.3.0",
    "tqdm>=4.66.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0"
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.26.0",
    "notebook>=7.0.0"
]

[project.scripts]
grid = "grid.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["grid*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
